External permission checks

+++ views/onboard.py
@@ -1,9 +1,11 @@
-from models import R2R, Role, Request, Onboard, db
+from models import R2R, Role, Request, Onboard, db, Staff
 from flask import render_template, session, redirect, url_for, flash
 from flask_login import current_user
 from forms import PersonForm, ApporovalForm
 from sqlalchemy import and_
 from . import models
+from . import permissions
+
 
 @models.route('/onboard')
 def onboard_list():
@@ -15,13 +17,16 @@ def onboard_list():
 
 @models.route('/onboard/select-r2r')
 def onboard_select_r2r():
-    r2rs = R2R.query.join(Request).join(Role).filter(and_(Request.status=='Approved', Role.school_id == session['active_school_id'])).all()
+    if current_user.is_admin:
+        r2rs = R2R.query.join(Request).join(Role).filter(Request.status=='Approved').all()
+    else:
+        r2rs = R2R.query.join(Request).join(Role).filter(and_(Request.status=='Approved', Role.school_id == session['active_school_id'])).all()
     return render_template('models/onboard/select_r2r.html', r2rs=r2rs)
 
 @models.route('/onboard/create/<int:r2r_id>', methods=['GET', 'POST'])
 def onboard_create(r2r_id):
     r2r = R2R.query.get_or_404(r2r_id)
-    if not perm_check_r2r(r2r):
+    if not permissions.onboard_create(r2r):
         return redirect(url_for('models.onboard_list'))
     pform = PersonForm()
     if pform.validate_on_submit():
@@ -38,20 +43,31 @@ def onboard_create(r2r_id):
 @models.route('/onboard/read/<int:onboard_id>', methods=['GET', 'POST'])
 def onboard_read(onboard_id):
     onboard = Onboard.query.get_or_404(onboard_id)
-    if not perm_check_onboard(onboard):
+    if not permissions.onboard_read(onboard):
         return redirect(url_for('models.onboard_list'))
-    aform = ApporovalForm()
+    aform = ApporovalForm(obj=onboard.request)
     if aform.validate_on_submit():
+        if not permissions.onboard_change(onboard):
+            return render_template('models/onboard/read.html', onboard=onboard, aform=aform)
+
+
         aform.populate_obj(onboard.request)
-        db.session.commit()
         flash(f'Status updated to {onboard.request.status}', 'success')
+        if onboard.request.status == 'Approved':
+            staff = Staff()
+            staff.role = onboard.r2r.role
+            for attribute in ['firstname','lastname','dob','gender','nino','nic','marital','home_address','postcode','email','startdate']:
+                setattr(staff, attribute, getattr(onboard, attribute))
+            db.session.add(staff)
+            flash(f"{staff} added to staff list", "success")
+        db.session.commit()
         return redirect(url_for('models.onboard_list'))
     return render_template('models/onboard/read.html', onboard=onboard, aform=aform)
 
 @models.route('/onboard/update/<int:onboard_id>', methods=['GET', 'POST'])
 def onboard_update(onboard_id):
     onboard = Onboard.query.get_or_404(onboard_id)
-    if not perm_check_onboard(onboard):
+    if not permissions.onboard_change(onboard):
         return redirect(url_for('models.onboard_list'))
     pform = PersonForm(obj=onboard)
     if pform.validate_on_submit():
@@ -62,27 +78,12 @@ def onboard_update(onboard_id):
 
 @models.route('/onboard/delete/<int:onboard_id>', methods=['POST'])
 def onboard_delete(onboard_id):
-    onboard = Onboard.query.get_or_404(onboard_id)
-    onboard.r2r.request.status = 'Approved'
+    onboard = db.session.get(Onboard,onboard_id)
+    if not permissions.onboard_change(onboard):
+        return redirect(url_for('models.onboard_list'))
     db.session.delete(onboard.request)
+    db.session.delete(onboard.role)
     db.session.delete(onboard)
     db.session.commit()
     return redirect(url_for('models.onboard_list'))
 
-def perm_check_r2r(r2r):
-    if not current_user.is_admin:
-        cond1 = r2r.role.school_id != session['active_school_id']
-        cond2 = r2r.request.status != 'Approved'
-        if cond1 or cond2:
-            flash('You do not have permissions to make changes here', 'error')
-            return False     
-    return True
-
-def perm_check_onboard(onboard):
-    if not current_user.is_admin:
-        cond1 = onboard.r2r.role.school_id != session['active_school_id']
-        cond2 = onboard.request.status != 'Pending'
-        if cond1 or cond2:
-            flash('You do not have permissions to make changes here', 'error')
-            return False     
-    return True
